Updated some JavaDoc,
Also added a check for 1 neighbour that is marked as a dead end, if so I have set it to back track.

Have also moved the getter methods for list to the top to make clearer.

Renamed the variable 'n' to 'neighbour' as it is easier to read and more descriptive